% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_division_as_choice_label}
\alias{filter_ui_division_as_choice_label}
\title{filter_ui_division_as_choice_label}
\usage{
filter_ui_division_as_choice_label(
  league,
  division,
  first_year = NULL,
  final_year = NULL,
  min_first_year = NULL,
  max_final_year = NULL
)
}
\arguments{
\item{league}{League}

\item{division}{Division (or \code{NA})}

\item{first_year}{First year the division is active}

\item{final_year}{Final year the division is active}

\item{min_first_year}{First year of the year range}

\item{max_final_year}{Final year of the year range}
}
\value{
League/Division/Year string
}
\description{
Given a league and division and, optionally, first and final years, as well
as a year range, produce strings like this:
\itemize{
\item "AL East"
\item "AL East (1990-1995)
\item "AL No Division"
\item "AL No Division (1990-1995)"
If the first and final year is contained entirely within the year range,
then just the league and division will be returned.  But if the min
or max year extends beyond the year range, then a range of years will
be appended.
}
}
