% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_division_as_choice_value}
\alias{filter_ui_division_as_choice_value}
\title{filter_ui_division_as_choice_value}
\usage{
filter_ui_division_as_choice_value(league, division)
}
\arguments{
\item{league}{League}

\item{division}{Division, or \code{NA}}
}
\value{
String of the form League_Division
}
\description{
Build a string like "AL_West" or "NL_None"
}
