% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_division_choice_value_as_league_and_division}
\alias{filter_ui_division_choice_value_as_league_and_division}
\title{filter_ui_division_choice_value_as_league_and_division}
\usage{
filter_ui_division_choice_value_as_league_and_division(choice_value)
}
\arguments{
\item{choice_value}{String of the form "XXX_YYY"}
}
\value{
list with \code{league} and \code{division}
}
\description{
Convert a string of the form "XXX_YYY" into a list of the form
\code{list(league="XXX", division="YYY")}.  If "YYY" is "None", then \code{division}
will be \code{NA}.
}
