% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_division_choice_values_as_league_and_division_list}
\alias{filter_ui_division_choice_values_as_league_and_division_list}
\title{filter_ui_division_choice_values_as_league_and_division_list}
\usage{
filter_ui_division_choice_values_as_league_and_division_list(choice_values)
}
\arguments{
\item{choice_values}{Vector of strings of the from "XXX_YYY"}
}
\value{
List of lists with \code{league} and \code{division}
}
\description{
Convert a vector of strings of the form "XXX_YYY" into a list of lists of the
form \code{list(league="XXX", division="YYY")}.  If "YYY" is "None", then
\code{division} will be \code{NA}.
}
