% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_filter_by_league}
\alias{filter_ui_filter_by_league}
\title{filter_ui_filter_by_league}
\usage{
filter_ui_filter_by_league(franchises, leagues)
}
\arguments{
\item{franchises}{Franchises table}

\item{leagues}{vector of leagues}
}
\value{
Filtered franchises table
}
\description{
Filter the franchises table by leagues
}
