% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_filter_by_league_division}
\alias{filter_ui_filter_by_league_division}
\title{filter_ui_filter_by_league_division}
\usage{
filter_ui_filter_by_league_division(franchises, league_division)
}
\arguments{
\item{franchises}{Franchises table}

\item{league_division}{League and division}
}
\value{
Filtered franchises table
}
\description{
Filter franchises table by league and division.  \code{league_division} should
be a list with \code{league} and \code{division}.
}
