% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_filter_by_years}
\alias{filter_ui_filter_by_years}
\title{filter_ui_filter_by_years}
\usage{
filter_ui_filter_by_years(franchises, years)
}
\arguments{
\item{franchises}{Franchises table}

\item{years}{vector of years.}
}
\value{
Filtered franchises table
}
\description{
Filter the franchises table by a set of years. Return franchises matching
any year in this list (but the list must consist of consecutive integers).
}
