% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_generate_league_division_selection}
\alias{filter_ui_generate_league_division_selection}
\title{filter_ui_generate_league_division_selection}
\usage{
filter_ui_generate_league_division_selection(division_table, league)
}
\arguments{
\item{division_table}{Division table}

\item{league}{League to filter by}
}
\value{
Named list
}
\description{
Given a division table and a league return a named list of choices, which
may look something like this:\preformatted{list("AL East (1969-1985)"="AL_East",
     "AL West (1980-2021)"="AL_West")
}
}
