% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_get_divisions}
\alias{filter_ui_get_divisions}
\title{filter_ui_get_divisions}
\usage{
filter_ui_get_divisions(franchises)
}
\arguments{
\item{franchises}{Franchise table}
}
\value{
Table with \code{League}, \code{Division}, \code{FirstSeason}, \code{FinalSeason}
}
\description{
Given a franchises table (which as already been truncated to to a
range of years), return all the league/division combos, as well
as the first and final seasons.
}
