% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_functions_ui.R
\name{filter_ui_truncate_years}
\alias{filter_ui_truncate_years}
\title{filter_ui_truncate_years}
\usage{
filter_ui_truncate_years(franchises, years)
}
\arguments{
\item{franchises}{Franchises table}

\item{years}{vector of years.}
}
\value{
Modified franchises table
}
\description{
Given a franchises table and a set of year, modify rows of the table so that
the \code{FirstSeason} and \code{FinalSeason} are 'truncated' by the min and max of the
years.  For example, if a row has \code{FinalSeason} \code{NA} or 2000, and the max
value in \code{years} is 1990, change the \code{FinalSeason} to 1990.
}
\details{
The assumption is that the \code{franchises} table has already been filtered by
the \code{years} list.
}
