% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{franchises_by_season}
\alias{franchises_by_season}
\title{franchises_by_season}
\usage{
franchises_by_season(franchises, year)
}
\arguments{
\item{franchises}{Franchises table}

\item{year}{Year}
}
\value{
Matching rows in the franchises table
}
\description{
Given the franchises table, find all the entries matching a specific year.
}
