% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{franchises_franchise_ids_to_team_ids}
\alias{franchises_franchise_ids_to_team_ids}
\title{franchises_franchise_ids_to_team_ids}
\usage{
franchises_franchise_ids_to_team_ids(
  franchises,
  franchise_ids,
  min_year,
  max_year
)
}
\arguments{
\item{franchises}{Franchise table}

\item{franchise_ids}{vector of FranchiseIDs}

\item{min_year}{First year}

\item{max_year}{Final year}
}
\value{
List of TeamIDs
}
\description{
Find all TeamIDs for the given FranchiseIDs and the year range.
}
