% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franchises.R
\name{franchises_get_division_by_team_year}
\alias{franchises_get_division_by_team_year}
\title{franchises_get_division_by_team_year}
\usage{
franchises_get_division_by_team_year(franchises, team, year)
}
\arguments{
\item{franchises}{Franchises table}

\item{team}{teamID}

\item{year}{year}
}
\value{
list with \code{division} and \code{teams}
}
\description{
Give a TeamID and year, find the division (\code{League}, \code{Division} and \code{Year})
and a table of teams in the division (\code{TeamID}, \code{Location}, \code{Nickname}).
}
