% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{lines_get_related_lines}
\alias{lines_get_related_lines}
\title{lines_get_related_lines}
\usage{
lines_get_related_lines(line_id, lzy_lines_to_streaks, lzy_concordances)
}
\arguments{
\item{line_id}{LineID}

\item{lzy_lines_to_streaks}{Lazy lines_to_streaks table}

\item{lzy_concordances}{Lazy concordances table}
}
\value{

}
\description{
Given a line_id, find the line_ids associated to all related streaks.
}
