% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{lines_get_selected_streak}
\alias{lines_get_selected_streak}
\title{lines_get_selected_streak}
\usage{
lines_get_selected_streak(
  lzy_lines_to_streaks,
  lzy_streaks,
  lzy_game_logs,
  line_id
)
}
\arguments{
\item{lzy_lines_to_streaks}{Lazy lines_to_streaks table}

\item{lzy_streaks}{Lazy streaks table}

\item{lzy_game_logs}{Lazy game_logs table}

\item{line_id}{LineID}
}
\value{
Streak
}
\description{
Given a line_id, get the related streak, as single-row table with
columns \code{Year}, \code{Team}, \code{LoIndex},\code{ HiIndex}, \code{StartDate}, and \code{EndDate}.
}
