% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaks.R
\name{pct_formatter}
\alias{pct_formatter}
\title{pct_formatter}
\usage{
pct_formatter(pct)
}
\arguments{
\item{pct}{Number to format}
}
\value{
Formatted winning percentage.
}
\description{
Formats winning percentages as a string with three digits after the
the decimal point and no leading zero, e.g. ".542".
Will return "1.000" for a perfect record.
}
