% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_add_lines_maybe}
\alias{plot_add_lines_maybe}
\title{plot_add_lines_maybe}
\usage{
plot_add_lines_maybe(p, lines = NULL, ...)
}
\arguments{
\item{p}{plot_ly object}

\item{lines}{Lines to add}

\item{...}{Additional params for plotly::add_lines}
}
\value{
modified plot_l object
}
\description{
Given plot_ly object and a set of table of lines, add the lines to
the object. Act as a passthrough if the lines table is omitted or
empty.
}
