% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lines}
\alias{plot_lines}
\title{plot_lines}
\usage{
plot_lines(
  lines,
  min_intensity,
  max_intensity,
  max_rank,
  highlighting,
  reverse_x_axis = FALSE
)
}
\arguments{
\item{lines}{Lines to plot}

\item{reverse_x_axis}{}
}
\value{
plotly::plot_ly object
}
\description{
Build the plotly plot.  The lines should have already been
passed through lines_highlight
}
