% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_standings_graph}
\alias{plot_standings_graph}
\title{plot_standings_graph}
\usage{
plot_standings_graph(standings, team, start_date, end_date)
}
\arguments{
\item{standings}{Standings to plot}

\item{team}{TeamID to highlight}

\item{start_date}{Start of highlight area}

\item{end_date}{End of highlight area}
}
\value{
Standings plot
}
\description{
Plot the standings graph, and highlight the selected team and date
range
}
