% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_functions_highlight_lines.R
\name{ps_get_related_lines}
\alias{ps_get_related_lines}
\title{ps_get_related_lines}
\usage{
ps_get_related_lines(line_id, lzy_lines_to_streaks, lzy_concordances)
}
\arguments{
\item{line_id}{LineID}

\item{lzy_lines_to_streaks}{Lazy lines_to_streaks table}

\item{lzy_concordances}{Lazy concordances table}
}
\value{
A vector of LineIDs
}
\description{
Given a line_id, find the line_ids associated to all related streaks.
}
