% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_functions_highlight_lines.R
\name{ps_lines_highlight}
\alias{ps_lines_highlight}
\title{ps_lines_highlight}
\usage{
ps_lines_highlight(
  lines,
  lzy_concordances,
  lzy_lines_to_streaks,
  line_id = NULL
)
}
\arguments{
\item{lines}{Table of lines}

\item{lzy_concordances}{Lazy concordances table}

\item{lzy_lines_to_streaks}{Lazy lines_to_streaks table}

\item{line_id}{LineID}
}
\value{
Lines with added \code{line_type} column.
}
\description{
Given a table of \code{lines} and a \code{line_id}, add a new column, \code{line_type},
with values:
\itemize{
\item \code{"identical"}: Matching line_id
\item \code{"related"}: Line represents a sub- or super-streak, but is not identical
\item \code{"season"}: Line is from same year and team, but not a sub- or super-streak
\item \code{"base"}: Unrelated line
}
}
\details{
If \code{line_id} is \code{NULL} or omitted, all values will be \code{"base"}
}
