% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_functions_server.R
\name{ps_plot_add_lines_maybe}
\alias{ps_plot_add_lines_maybe}
\title{ps_plot_add_lines_maybe}
\usage{
ps_plot_add_lines_maybe(p, lines = NULL, ...)
}
\arguments{
\item{p}{plot_ly object}

\item{lines}{Lines to add}

\item{...}{Additional params for plotly::add_lines}
}
\value{
modified plot_ly object
}
\description{
Given plot_ly object and a set of table of lines, add the lines to
the object. Act as a passthrough if the lines table is omitted or
empty.
}
