% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_functions_server.R
\name{ps_plot_lines}
\alias{ps_plot_lines}
\title{ps_plot_lines}
\usage{
ps_plot_lines(
  lines,
  min_intensity,
  max_intensity,
  max_rank,
  highlighting,
  reverse_x_axis = FALSE
)
}
\arguments{
\item{lines}{Lines to plot}

\item{min_intensity}{Min intensity level}

\item{max_intensity}{Max intensity level}

\item{max_rank}{Max rank}

\item{highlighting}{Highlighting info (line colors and widths)}

\item{reverse_x_axis}{If TRUE, reverse the x-axis}
}
\value{
plotly::plot_ly object
}
\description{
Build the plotly plot.  The lines should have already been
passed through lines_highlight
}
