% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_functions_server.R
\name{ps_streaks_get_max_rank}
\alias{ps_streaks_get_max_rank}
\title{ps_streaks_get_max_rank}
\usage{
ps_streaks_get_max_rank(n, min_year, max_year, teams, hot)
}
\arguments{
\item{n}{Function will maximize value of \code{n}th highest rank}

\item{min_year}{Minimum year for filter}

\item{max_year}{Maximum year for filter}

\item{teams}{Vector of team IDs for filter.}

\item{hot}{If \code{TRUE} use hot streaks, otherwise cold streaks}
}
\value{
Max value estimate
}
\description{
Given a year range and a list of teams and a value n, find an
estimate of the maximum of the nth highest rank over all intensity levels.
}
\details{
This function is a wrapper around \code{ps_streaks_get_max_rank_simple} and
\code{ps_streaks_get_max_rank_by_sampling} and will decide the appropriate
function and parameters to use based on the number of years and
the number of teams.
}
