% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_server.R
\name{ss_build_standings_graph}
\alias{ss_build_standings_graph}
\title{ss_build_standings_graph}
\usage{
ss_build_standings_graph(lzy_standings, franchises, streak)
}
\arguments{
\item{lzy_standings}{Lazy full standings table}

\item{franchises}{Franchises table}

\item{streak}{Streak info}
}
\value{
Standings plot
}
\description{
Given the streak info, build the standings table for the division
and pass it to ss_plot_standings_graph to generate a standings plot
}
