% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_server.R
\name{ss_franchises_by_season}
\alias{ss_franchises_by_season}
\title{ss_franchises_by_season}
\usage{
ss_franchises_by_season(franchises, year)
}
\arguments{
\item{franchises}{Franchises table}

\item{year}{Year}
}
\value{
Matching rows in the franchises table
}
\description{
Given the franchises table, find all the entries matching a specific year.
}
