% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_server.R
\name{ss_get_same_day_team_games_lzy}
\alias{ss_get_same_day_team_games_lzy}
\title{ss_get_same_day_team_games_lzy}
\usage{
ss_get_same_day_team_games_lzy(
  lzy_game_logs,
  year,
  season_game_id,
  before = TRUE
)
}
\arguments{
\item{lzy_game_logs}{Lazy game logs table}

\item{year}{Year}

\item{season_game_id}{SeasonGameId}

\item{before}{If \code{TRUE}, game games earlier in the day. If \code{FALSE}, later}
}
\value{
Lazy query
}
\description{
Given the year and the SeasonGameId, find all games between the same two
teams either earlier in the same day (when \code{before} is \code{TRUE}) or later
in the same day (when \code{before} is \code{FALSE}).  The result is a lazy query
to the game_logs table.
}
\details{
The result is normally empty, and is only interesting when there is a
double-header.
}
