% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_server.R
\name{ss_pct_formatter}
\alias{ss_pct_formatter}
\title{ss_pct_formatter}
\usage{
ss_pct_formatter(pct)
}
\arguments{
\item{pct}{Number to format}
}
\value{
Formatted winning percentage.
}
\description{
Formats winning percentages as a string with three digits after the
the decimal point and no leading zero, e.g. ".542".
Will return "1.000" for a perfect record.
}
