% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_server.R
\name{ss_plot_standings_graph}
\alias{ss_plot_standings_graph}
\title{ss_plot_standings_graph}
\usage{
ss_plot_standings_graph(
  highlight_colors,
  standings,
  team,
  start_date,
  end_date
)
}
\arguments{
\item{highlight_colors}{Colors for highlighting}

\item{standings}{Standings to plot}

\item{team}{TeamID to highlight}

\item{start_date}{Start of highlight area}

\item{end_date}{End of highlight area}
}
\value{
Standings plot
}
\description{
Plot the standings graph, and highlight the selected team and date
range
}
