% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_server.R
\name{ss_streak_game_log_data}
\alias{ss_streak_game_log_data}
\title{ss_streak_game_log_data}
\usage{
ss_streak_game_log_data(lzy_game_logs, streak)
}
\arguments{
\item{lzy_game_logs}{Lazy game logs}

\item{streak}{streak}
}
\value{
list with \code{data} and \code{caption}
}
\description{
Given a streak -- a list with fields Year, Team, LoIndex and HiIndex --
return a list with two elements: \code{caption}
which contains a string like "Game Log", and \code{data}, which is a table
suitable for displaying as a game log
}
