% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_server.R
\name{ss_streak_summary_data}
\alias{ss_streak_summary_data}
\title{ss_streak_summary_data}
\usage{
ss_streak_summary_data(lzy_game_logs, franchises, streak)
}
\arguments{
\item{lzy_game_logs}{Lazy game logs table}

\item{franchises}{franchise table}

\item{streak}{streak}
}
\value{
list with \code{data} and \code{caption}
}
\description{
Build streak summary data from a streak.  The will be two data elements
to the return values:  \code{data} will be a single-row summary table suitable
for displaying, and \code{caption} will contain a string like
"2004 Cincinnati Reds, Games 100-104"
}
