% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_server.R
\name{ss_streaks_get_game_log}
\alias{ss_streaks_get_game_log}
\title{ss_streaks_get_game_log}
\usage{
ss_streaks_get_game_log(lzy_game_logs, streak)
}
\arguments{
\item{lzy_game_logs}{Lazy game log}

\item{streak}{Streak}
}
\value{
filtered game log
}
\description{
Filter a game log using a streak (using year, team, and the game index).
The input is lazy, and the output is in memory.
}
