% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_functions_standings.R
\name{ss_update_standings_from_game_logs}
\alias{ss_update_standings_from_game_logs}
\title{ss_update_standings_from_game_logs}
\usage{
ss_update_standings_from_game_logs(lzy_standings, lzy_games_to_add)
}
\arguments{
\item{lzy_standings}{Lazy standings table}

\item{lzy_games_to_add}{Game logs to add}
}
\value{
Updated standings
}
\description{
Add the results of the game logs to a standings table.
}
\details{
This is used for handling the case when we want to add a partial days
games to a precalculated standings table, because of a double-header.
}
