% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{standings_get_by_season_game_id}
\alias{standings_get_by_season_game_id}
\title{standings_get_by_season_game_id}
\usage{
standings_get_by_season_game_id(
  lzy_standings,
  lzy_game_logs,
  division,
  season_game_id,
  before = TRUE
)
}
\arguments{
\item{lzy_standings}{Lazy standings table}

\item{lzy_game_logs}{Lazy game logs table}

\item{division}{division table (\code{Year}, \code{League} and \code{Division})}

\item{season_game_id}{SeasonGameId}

\item{before}{If \code{TRUE}, use start-of-day. Otherwise, end-of-day.}
}
\value{
Standings table
}
\description{
Given a division and a SeasonGameId, get the standings, either at
the start of day (\code{before} is \code{TRUE}) or the end of day (\code{before} is
\code{FALSE}).
}
