% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{standings_get_final_standings}
\alias{standings_get_final_standings}
\title{standings_get_final_standings}
\usage{
standings_get_final_standings(lzy_standings, division)
}
\arguments{
\item{lzy_standings}{Lazy standings table}

\item{division}{Division}
}
\value{
Standings table.
}
\description{
Get the final standings for a division.  \code{division} should be a single-row
table with \code{Year}, \code{League} and \code{Division} (possibly \code{NA}).
}
