% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaks.R
\name{streaks_game_log_data}
\alias{streaks_game_log_data}
\title{streaks_game_log_data}
\usage{
streaks_game_log_data(lzy_game_logs, streak)
}
\arguments{
\item{streak}{}
}
\value{

}
\description{
Given a streak -- a list with fields Year, Team, LoIndex and HiIndex --
return a list with with elements, return a list with two elements: \code{caption}
which contains a string like "Game Log", and \code{data}, which is a table
suitable for displaying as a game log
}
