% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_util.R
\name{streaks_get_intensity_range}
\alias{streaks_get_intensity_range}
\title{streaks_get_intensity_range}
\usage{
streaks_get_intensity_range(lzy_streaks, year)
}
\arguments{
\item{lzy_streaks}{Lazy streaks table}

\item{year}{Year to query.}
}
\value{
Vector with min and max intensity levels, e.g. \code{c(1,101)}
}
\description{
Use the streaks table to get the the minimum and maximum intensity
levels in the DB.  The year is specified to make the query quicker.
}
