% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaks.R
\name{streaks_get_max_rank_simple}
\alias{streaks_get_max_rank_simple}
\title{streaks_get_max_rank_simple}
\usage{
streaks_get_max_rank_simple(lzy_streaks, n, min_year, max_year, teams)
}
\arguments{
\item{lzy_streaks}{Lazy streaks table}

\item{n}{Function will maximize value of \code{n}th highest rank}

\item{min_year}{Minimun year for filter}

\item{max_year}{Maximum year for filter}

\item{teams}{Vector of team IDs for filter.}
}
\value{
Maximum value
}
\description{
Given a year range and a list of teams and a value n, find the maximum
}
