% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaks.R
\name{streaks_get_related_streak_ids}
\alias{streaks_get_related_streak_ids}
\title{streaks_get_related_streak_ids}
\usage{
streaks_get_related_streak_ids(streak_id, lzy_concordances)
}
\arguments{
\item{streak_id}{Streak ID}

\item{lzy_concordances}{Lazy conconcrdance table}
}
\value{
vector of related streak IDs.
}
\description{
Given a streak ID, returns all related streaks in the concordance table --
both super-streaks and sub-streaks -- including the streak itself.
}
