% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaks.R
\name{streaks_get_standings}
\alias{streaks_get_standings}
\title{streaks_get_standings}
\usage{
streaks_get_standings(lzy_standings, lzy_game_logs, franchises, streak)
}
\arguments{
\item{lzy_standings}{Lazy standings table}

\item{lzy_game_logs}{Lazy game logs table}

\item{franchises}{Franchises table}

\item{streak}{Streak}
}
\value{
list with \code{streak_info}, \code{standings_before}, \code{standings_after},
and \code{standings_final}
}
\description{
Givn a streak, return the division standings before the streak started
(\code{standings_before}), after the last game of the streak (\code{standings_after}),
and at the end of the season (\code{standings_final}). Also compute a summary
object (\code{streak_info}).
}
