% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaks.R
\name{streaks_summary_data}
\alias{streaks_summary_data}
\title{streaks_summary_data}
\usage{
streaks_summary_data(lzy_game_logs, lzy_franchises, streak)
}
\arguments{
\item{lzy_game_logs}{Lazy game logs table}

\item{lzy_franchises}{Lazy franchise table}

\item{streak}{streak}
}
\value{
list with \code{data} and \code{caption}
}
\description{
Build streak summary data from a streak.  The will be two data elements
to the return values:  \code{data} will be a single-row summary table suitable
for displaying, and \code{caption} will contain a string like
"2004 Cincinnati Reds, Games 100-104"
}
