% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_util.R
\name{tbl_as_named_list}
\alias{tbl_as_named_list}
\title{Given a table and two columns, build a named list.}
\usage{
tbl_as_named_list(tbl, value_col, name_col)
}
\arguments{
\item{tbl}{Table}

\item{value_col}{Column with list values}

\item{name_col}{Column with list names}
}
\value{
A named list
}
\description{
Build a named list from the \code{value_col} and \code{name_col}. Note
that there is no uniqueness test on the names column.
}
