% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_util.R
\name{tbl_is_column_value_unique}
\alias{tbl_is_column_value_unique}
\title{Check that a column has a unique value}
\usage{
tbl_is_column_value_unique(tbl, column_name)
}
\arguments{
\item{tbl}{Data table}

\item{column_name}{Column name (unquoted)}
}
\value{
TRUE or FALSE
}
\description{
Check that a column has a unique value
}
