% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_util.R
\name{ui_build_divisions_choices}
\alias{ui_build_divisions_choices}
\title{ui_build_divisions_choices}
\usage{
ui_build_divisions_choices(franchises, years, leagues)
}
\arguments{
\item{franchises}{Franchise table (unfiltered)}

\item{years}{Vector of years to filter on}

\item{leagues}{League filter: "AL", "NL", or c("AL","NL")}
}
\value{
List of choices for the UI
}
\description{
Given a division table and a year and league filter,
generate a choice list for the division selection
UI.  If all the divisions are from the same league, there will be an
element for each division. If both leagues are represented, then there will
be two sublists, labeled "AL Divisions" and "NL Divisions".  (This will
generate separators in the UI)
}
