% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_util.R
\name{ui_build_teams_choices}
\alias{ui_build_teams_choices}
\title{ui_build_teams_choices}
\usage{
ui_build_teams_choices(franchises, years, league_divisions)
}
\arguments{
\item{franchises}{Franchises table}

\item{years}{Years filter}

\item{league_divisions}{List of lists with \code{league} and \code{division}}
}
\value{
Named list for the UI
}
\description{
Given the franchises table, and a years and league/division filter,
generate a named list of FranchiseIDs, where the names of the
franchises is based on the teams Nicknames, like "Red Sox" or
"Rays/Devil Rays".
}
