% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_util.R
\name{ui_filter_by_league}
\alias{ui_filter_by_league}
\title{ui_filter_by_league}
\usage{
ui_filter_by_league(franchises, leagues)
}
\arguments{
\item{franchises}{Franchises table}

\item{leagues}{vector of leagues}
}
\value{
Filtered franchises table
}
\description{
Filter the franchises table by leagues
}
