% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_util.R
\name{ui_filter_by_league_divisions}
\alias{ui_filter_by_league_divisions}
\title{ui_filter_by_league_divisions}
\usage{
ui_filter_by_league_divisions(franchises, league_divisions)
}
\arguments{
\item{franchises}{Franchises table}

\item{league_divisions}{List of lists}
}
\value{
Filtered franchises table
}
\description{
Filter franchises table by a list of leagues and divisions.
\code{league_divisions} should be a list of lists, each with \code{league} and
\code{division}.
}
