% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_util.R
\name{ui_filter_by_year}
\alias{ui_filter_by_year}
\title{ui_filter_by_year}
\usage{
ui_filter_by_year(franchises, year)
}
\arguments{
\item{franchises}{Franchises table}

\item{year}{Year}
}
\value{
Filtered franchises table
}
\description{
Filter the franchises table by year.
}
