% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_util.R
\name{ui_filter_by_years}
\alias{ui_filter_by_years}
\title{ui_filter_by_years}
\usage{
ui_filter_by_years(franchises, years)
}
\arguments{
\item{franchises}{Franchises table}

\item{years}{vector of years.}
}
\value{
Filtered franchises table
}
\description{
Filter the franchises table by a set of years. Return franchises matching
any year in this list.
}
