% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_util.R
\name{ui_generate_division_selection}
\alias{ui_generate_division_selection}
\title{ui_generate_division_selection}
\usage{
ui_generate_division_selection(division_table)
}
\arguments{
\item{division_table}{Division table}
}
\value{
List of choices for the UI
}
\description{
Given a division table, generate a choice list for the division selection
UI.  If all the divisions are from the same league, there will be an
element for each division. If both leagues are represented, then there will
be two sublists, labeled "AL Divisions" and "NL Divisions".  (This will
generate separators in the UI)
}
