% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_util.R
\name{ui_generate_team_selection}
\alias{ui_generate_team_selection}
\title{ui_generate_team_selection}
\usage{
ui_generate_team_selection(franchises)
}
\arguments{
\item{franchises}{Franchises table}
}
\value{
Named list of franchise IDs.
}
\description{
Given a (filtered) franchises table, generate a named list of teams, with
the distinct FranchiseIDs as values.  The name of each value will be a
combination of all nicknames associated to the FranchiseID -- e.g.
"Orioles/Browns"="BAL".  For each franchise, arrange it so the the nicknames
are arranged from most recent to least recent.
}
